# apiary-OpenApi3.yaml
openapi: 3.0.0
info:
  title: Financials Microservice
  description: This is the documentation of the Financials Microservice for the Soaring the Cloud Sequel Showcase.
  contact: {}
  version: '1.0'
servers:
- url: http://private-0b1f3-financialsmicroservice.apiary-mock.com/financials/api/v1
  variables: {}
paths:
  /invoices:
    get:
      tags:
      - InvoicesResources
      summary: ApiFinancialsInvoicesGet
      description: Search invoices using criterias like the Customer ID or Order ID
      operationId: ApiFinancialsInvoicesGet
      parameters:
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: customer_id
        in: query
        description: Filter by customer_id
        style: form
        explode: true
        schema:
          type: string
      - name: order_id
        in: query
        description: Filter by order_id
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
              example:
              - invoice_id: I2134354535
                order_id: order4711
                order_date: '"2018-02-17T21:00:00Z"'
                shipping_costs: 6.9900000000000002
                total_price: 31.960000000000001
                currency: EUR
                invoice_status: OPEN
                payment_status: PENDING
                customer:
                  customer_no: CGN4711
                  first_name: Sven
                  last_name: Bernhardt
                  phone: +49 211 5093733
                  email: sven@yourmail.com
                addresses:
                - type: BILLING
                  street_number: 15
                  street: Ludwig-Jahn-Strasse
                  city: Köln
                  postcode: 50858
                  country: DE
                invoice_positions:
                - position_id: 1
                  amount: 1
                  description: 'Blue-Ray Box: Game of Thrones Season 7'
                  product_number: B073ZKVHFV
                  net_price: 20.98
                  vat: 19
                  gross_price: 24.969999999999999
      deprecated: false
    post:
      tags:
      - InvoicesResources
      summary: ApiFinancialsInvoicesPost
      description: Create a new invoice
      operationId: ApiFinancialsInvoicesPost
      parameters:
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: customer_id
        in: query
        description: Filter by customer_id
        style: form
        explode: true
        schema:
          type: string
      - name: order_id
        in: query
        description: Filter by order_id
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
        required: true
      responses:
        201:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMetadata'
              example:
                error: true
                processing_message: Invoice created successfully!
      deprecated: false
  /invoices/{invoice_id}:
    get:
      tags:
      - InvoicesResources
      summary: ApiFinancialsInvoicesByInvoiceIdGet
      description: get a specific invoice by it's Invoice ID
      operationId: ApiFinancialsInvoicesByInvoiceIdGet
      parameters:
      - name: invoice_id
        in: path
        description: invoice_id of invoice to manage
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
              example:
                invoice_id: I2134354535
                order_id: order4711
                order_date: '"2018-02-17T21:00:00Z"'
                shipping_costs: 6.9900000000000002
                total_price: 31.960000000000001
                currency: EUR
                invoice_status: OPEN
                payment_status: PENDING
                customer:
                  customer_no: CGN4711
                  first_name: Sven
                  last_name: Bernhardt
                  phone: +49 211 5093733
                  email: sven@yourmail.com
                addresses:
                - type: BILLING
                  street_number: 15
                  street: Ludwig-Jahn-Strasse
                  city: Köln
                  postcode: 50858
                  country: DE
                invoice_positions:
                - position_id: 1
                  amount: 1
                  description: 'Blue-Ray Box: Game of Thrones Season 7'
                  product_number: B073ZKVHFV
                  net_price: 20.98
                  vat: 19
                  gross_price: 24.969999999999999
      deprecated: false
    put:
      tags:
      - InvoicesResources
      summary: ApiFinancialsInvoicesByInvoiceIdPut
      description: Update a specific invoice
      operationId: ApiFinancialsInvoicesByInvoiceIdPut
      parameters:
      - name: invoice_id
        in: path
        description: invoice_id of invoice to manage
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
        required: true
      responses:
        200:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMetadata'
              example:
                error: true
                processing_message: Invoice created successfully!
      deprecated: false
    delete:
      tags:
      - InvoicesResources
      summary: ApiFinancialsInvoicesByInvoiceIdDelete
      description: 'Deletes a specific invoice (NOTE: FOR INTERNAL USE ONLY! Use carefully!)'
      operationId: ApiFinancialsInvoicesByInvoiceIdDelete
      parameters:
      - name: invoice_id
        in: path
        description: invoice_id of invoice to manage
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: ''
          headers: {}
      deprecated: false
  /payments:
    get:
      tags:
      - PaymentsResources
      summary: ApiFinancialsPaymentsGet
      description: Get the relevant payment information for an invoice
      operationId: ApiFinancialsPaymentsGet
      parameters:
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: invoice_id
        in: query
        description: Filter for invoice_id
        style: form
        explode: true
        schema:
          type: string
      - name: customer_no
        in: query
        description: Filter for customer_no
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
              example:
              - payment_id: pay666
                invoice_id: I2134354535
                order_id: order4711
                customer_no: CGN4711
                payment_type: VISA
                payment_state: PENDING
      deprecated: false
    post:
      tags:
      - PaymentsResources
      summary: ApiFinancialsPaymentsPost
      description: Create a new payment
      operationId: ApiFinancialsPaymentsPost
      parameters:
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: invoice_id
        in: query
        description: Filter for invoice_id
        style: form
        explode: true
        schema:
          type: string
      - name: customer_no
        in: query
        description: Filter for customer_no
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
        required: true
      responses:
        201:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMetadata'
              example:
                error: true
                processing_message: Invoice created successfully!
      deprecated: false
  /customers/{customer_id}/account:
    get:
      tags:
      - CustomerAccountResources
      summary: ApiFinancialsCustomersAccountByCustomerIdGet
      description: Getcustomeraccountinfo
      operationId: ApiFinancialsCustomersAccountByCustomerIdGet
      parameters:
      - name: customer_id
        in: path
        description: customer_id of customer to manage account for
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: Content-Type
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAccount'
              example:
                customer_no: CGN4711
                first_name: Sven
                last_name: Bernhardt
                balance: -69.989999999999995
                currency: EUR
                customer_status: FROZEN
      deprecated: false
  /health:
    get:
      tags:
      - ServiceOperationsResources
      summary: ApiFinancialsHealthGet
      description: Get service health information
      operationId: ApiFinancialsHealthGet
      parameters: []
      responses:
        200:
          description: ''
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
              example:
                version: 0.10000000000000001
                status: OK
                uptime: 1234.0999999999999
      deprecated: false
components:
  schemas:
    HealthCheckResponse:
      title: HealthCheckResponse
      required:
      - status
      type: object
      properties:
        version:
          type: number
          example: 0.10000000000000001
        status:
          $ref: '#/components/schemas/status'
        uptime:
          type: number
          description: The uptime of the service in minutes
          example: 1234.0999999999999
      example:
        version: 0.10000000000000001
        status: OK
        uptime: 1234.0999999999999
    status:
      title: status
      enum:
      - OK
      - FAILED
      type: string
      example: OK
    ResponseMetadata:
      title: ResponseMetadata
      required:
      - error
      - processing_message
      type: object
      properties:
        error:
          type: boolean
          example: true
        processing_message:
          type: string
          example: Invoice created successfully!
      example:
        error: true
        processing_message: Invoice created successfully!
    Customer:
      title: Customer
      required:
      - customer_no
      - first_name
      - last_name
      - email
      type: object
      properties:
        customer_no:
          type: string
          example: CGN4711
        first_name:
          type: string
          example: Sven
        last_name:
          type: string
          example: Bernhardt
        phone:
          type: string
          example: +49 211 5093733
        email:
          type: string
          example: sven@yourmail.com
    Address:
      title: Address
      required:
      - type
      - street_number
      - city
      - postcode
      - country
      type: object
      properties:
        type:
          $ref: '#/components/schemas/type'
        street_number:
          type: string
          example: 15
        street:
          type: string
          example: Ludwig-Jahn-Strasse
        city:
          type: string
          example: Köln
        postcode:
          type: string
          example: 50858
        country:
          type: string
          example: DE
    type:
      title: type
      enum:
      - BILLING
      - DELIVERY
      type: string
      example: BILLING
    InvoicePosition:
      title: InvoicePosition
      required:
      - position_id
      - amount
      - description
      - product_number
      - net_price
      - vat
      - gross_price
      type: object
      properties:
        position_id:
          type: number
          example: 1
        amount:
          type: number
          example: 1
        description:
          type: string
          example: 'Blue-Ray Box: Game of Thrones Season 7'
        product_number:
          type: string
          example: B073ZKVHFV
        net_price:
          type: number
          example: 20.98
        vat:
          type: number
          example: 19
        gross_price:
          type: number
          example: 24.969999999999999
    Invoice:
      title: Invoice
      required:
      - invoice_id
      - order_id
      - shipping_costs
      - total_price
      type: object
      properties:
        invoice_id:
          type: string
          description: System generated Invoice ID
          example: I2134354535
        order_id:
          type: string
          example: order4711
        order_date:
          type: string
          example: 2/17/2018 9:00:00 PM
        shipping_costs:
          type: number
          example: 6.9900000000000002
        total_price:
          type: number
          example: 31.960000000000001
        currency:
          $ref: '#/components/schemas/currency'
        invoice_status:
          $ref: '#/components/schemas/invoice_status'
        payment_status:
          $ref: '#/components/schemas/payment_status'
        customer:
          $ref: '#/components/schemas/Customer'
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
          description: ''
        invoice_positions:
          type: array
          items:
            $ref: '#/components/schemas/InvoicePosition'
          description: ''
      example:
        invoice_id: I2134354535
        order_id: order4711
        order_date: '"2018-02-17T21:00:00Z"'
        shipping_costs: 6.9900000000000002
        total_price: 31.960000000000001
        currency: EUR
        invoice_status: OPEN
        payment_status: PENDING
        customer:
          customer_no: CGN4711
          first_name: Sven
          last_name: Bernhardt
          phone: +49 211 5093733
          email: sven@yourmail.com
        addresses:
        - type: BILLING
          street_number: 15
          street: Ludwig-Jahn-Strasse
          city: Köln
          postcode: 50858
          country: DE
        invoice_positions:
        - position_id: 1
          amount: 1
          description: 'Blue-Ray Box: Game of Thrones Season 7'
          product_number: B073ZKVHFV
          net_price: 20.98
          vat: 19
          gross_price: 24.969999999999999
    currency:
      title: currency
      enum:
      - EUR
      - GBP
      - USD
      type: string
      example: EUR
    invoice_status:
      title: invoice_status
      enum:
      - OPEN
      - PAID
      - CANCEL
      type: string
      example: OPEN
    payment_status:
      title: payment_status
      enum:
      - PENDING
      - RECEIVED
      type: string
      example: PENDING
    Payment:
      title: Payment
      required:
      - invoice_id
      - order_id
      - customer_no
      - payment_type
      type: object
      properties:
        payment_id:
          type: string
          example: pay666
        invoice_id:
          type: string
          example: I2134354535
        order_id:
          type: string
          example: order4711
        customer_no:
          type: string
          example: CGN4711
        payment_type:
          $ref: '#/components/schemas/payment_type'
        payment_state:
          $ref: '#/components/schemas/payment_state'
      example:
        payment_id: pay666
        invoice_id: I2134354535
        order_id: order4711
        customer_no: CGN4711
        payment_type: VISA
        payment_state: PENDING
    payment_type:
      title: payment_type
      enum:
      - VISA
      - Mastercard
      - AMEX
      - PayPal
      - Invoice
      type: string
      example: VISA
    payment_state:
      title: payment_state
      enum:
      - PENDING
      - APPROVED
      type: string
      example: PENDING
    CustomerAccount:
      title: CustomerAccount
      required:
      - customer_no
      - first_name
      - last_name
      - balance
      - customer_status
      type: object
      properties:
        customer_no:
          type: string
          example: CGN4711
        first_name:
          type: string
          example: Sven
        last_name:
          type: string
          example: Bernhardt
        balance:
          type: number
          example: -69.989999999999995
        currency:
          $ref: '#/components/schemas/currency'
        customer_status:
          $ref: '#/components/schemas/customer_status'
      example:
        customer_no: CGN4711
        first_name: Sven
        last_name: Bernhardt
        balance: -69.989999999999995
        currency: EUR
        customer_status: FROZEN
    customer_status:
      title: customer_status
      enum:
      - FROZEN
      - NORMAL
      type: string
      example: FROZEN
tags:
- name: Invoices Resources
  description: ''
- name: Payments Resources
  description: ''
- name: Customer Account Resources
  description: ''
- name: Service Operations Resources
  description: ''
- name: InvoicesResources
  description: ''
- name: PaymentsResources
  description: ''
- name: CustomerAccountResources
  description: ''
- name: ServiceOperationsResources
  description: ''